refactor: widen client config dimensions and guard SDL size cast

+++ src/client/client_backend.h
@@ -6,8 +6,8 @@
 
 typedef struct bc_client_config_s {
     const char *title;
-    uint16_t width;
-    uint16_t height;
+    uint32_t width;
+    uint32_t height;
 } bc_client_config_t;
 
 /*
